Added memorypool usage and removed a lot of allocations

+++ Makefile
@@ -1,5 +1,6 @@
 #Shell
 SHELL = /bin/bash 
+
 #Executeable name
 NAME = WSServer
 
@@ -10,56 +11,6 @@ CC = gcc
 ifeq ($(VER),)
 	VER = $(shell git describe --abbrev=0 --tags)
 endif
-
-#Debug or Release
-PROFILE = -Og -g -DNDEBUG
-DEBUG = -Og -g
-RELEASE = -O3 -funroll-loops -DNDEBUG
-SPACE = -Os -DNDEBUG
-MODE = "release"
-EXEC = $(RELEASE)
-
-#Compiler options
-CFLAGS = $(EXEC) \
-		 -fno-exceptions \
-		 -fPIC \
-		 -fstack-protector \
-		 -fvisibility=hidden \
-		 -march=native \
-		 -MMD \
-		 -pedantic \
-		 -pedantic-errors \
-		 -pipe \
-		 -W \
-		 -Wall \
-		 -Werror \
-		 -Wno-unused-parameter \
-		 -Wno-unused-function \
-		 -Wno-strict-aliasing \
-		 -Wno-overlength-strings \
-		 -Wformat \
-		 -Wformat-security \
-		 -Wformat-nonliteral \
-		 -Winit-self \
-		 -Winline \
-		 -Wl,-z,relro \
-		 -Wl,-z,now \
-		 -Wmultichar \
-		 -Wpointer-arith \
-		 -Wreturn-type \
-	     -Wsign-compare \
-		 -Wuninitialized \
-		 -D_DEFAULT_SOURCE \
-		 -DWSS_SERVER_VERSION=\"$(VER)\" \
-		 -DUSE_RPMALLOC #\
-		 -DUSE_POLL\
-
-CVER = -std=c11
-
-# Flags
-FLAGS_EXTRA = -pthread -lm -ldl
-FLAGS_TEST = -lgcov
-FLAGS_COVERAGE =
 
 # Folders
 ROOT = $(shell pwd)
@@ -76,36 +27,86 @@ EXTENSIONS_FOLDER = $(ROOT)/extensions
 SUBPROTOCOLS_FOLDER = $(ROOT)/subprotocols
 SCRIPTS_FOLDER = $(ROOT)/scripts
 
-# Include folders
-INCLUDES = -I$(INCLUDE_FOLDER) -I$(SRC_FOLDER) -I$(EXTENSIONS_FOLDER) -I$(SUBPROTOCOLS_FOLDER)
+#Debug or Release
+PROFILING = -Og -g -DNDEBUG
+DEBUG = -Og -g --param max-inline-insns-single=1000
+RELEASE = -O3 -funroll-loops -DNDEBUG
+SPACE = -Os -DNDEBUG
+MODE = "release"
+PREV_MODE = "$(shell cat $(BUILD_FOLDER)/.mode 2> /dev/null)"
+PREV_SSL = $(shell cat $(BUILD_FOLDER)/.ssl 2> /dev/null)
+EXEC = $(RELEASE)
+
+#Compiler options
+CFLAGS = $(EXEC) \
+		 -pedantic \
+		 -pedantic-errors \
+		 -fno-exceptions \
+		 -fPIC \
+		 -fstack-protector -Wl,-z,relro -Wl,-z,now \
+		 -fvisibility=hidden \
+		 -march=native \
+		 -MMD \
+		 -pipe \
+		 -W \
+		 -Wall \
+		 -Werror \
+		 -Wextra \
+		 -Wformat \
+		 -Wformat-security \
+		 -Wformat-nonliteral \
+		 -Winit-self \
+		 -Winline \
+		 -Wmultichar \
+		 -Wpointer-arith \
+		 -Wreturn-type \
+	     -Wsign-compare \
+		 -Wuninitialized \
+		 -D_GNU_SOURCE \
+		 -D_DEFAULT_SOURCE \
+		 -DWSS_SERVER_VERSION=\"$(VER)\" \
+		 -DUSE_RPMALLOC #\
+		 -DUSE_POLL\
+
+# C version
+CVER = -std=c11
+
+# Flags
+FLAGS_EXTRA = -pthread -lm -ldl
+FLAGS_TEST = -lgcov
+FLAGS_COVERAGE =
+FLAGS_INCLUDES = -I$(INCLUDE_FOLDER) -I$(SRC_FOLDER) -I$(EXTENSIONS_FOLDER) -I$(SUBPROTOCOLS_FOLDER)
 
 # Files
-SRC = $(shell find $(SRC_FOLDER) -name '*.c' -type f;)
-TESTS = $(shell find $(TEST_FOLDER) -name 'test_*.c' -type f;)
+SRC = $(shell find $(SRC_FOLDER) -iname '*.c' -type f;)
+TESTS = $(shell find $(TEST_FOLDER) -iname 'test_*.c' -type f;)
 SRC_OBJ  = $(subst $(SRC_FOLDER), $(BUILD_FOLDER), $(patsubst %.c, %.o, $(SRC)))
 TEST_OBJ = ${subst ${TEST_FOLDER}, ${BUILD_FOLDER}, ${patsubst %.c, %.o, $(TESTS)}}
 ALL_OBJ  = ${SRC_OBJ} ${TEST_OBJ}
 TEST_NAMES = ${patsubst ${TEST_FOLDER}/%.c, %, ${TESTS}}
 DEPS = $(ALL_OBJ:%.o=%.d)
 
-ifneq ($(SSL_LIBRARY_PATH),)
-	INCLUDES += -I$(SSL_LIBRARY_PATH)/include -L$(SSL_LIBRARY_PATH)/lib
-	export LD_LIBRARY_PATH=$(SSL_LIBRARY_PATH)/lib
-endif
-
 ifeq ($(BUMP),)
 	BUMP = default
+endif
+
+ifneq ($(SSL_LIBRARY_PATH),)
+	SSL_MODE=$(SSL_LIBRARY_PATH)
+	FLAGS_INCLUDES += -I$(SSL_LIBRARY_PATH)/include -L$(SSL_LIBRARY_PATH)/lib
+	export LD_LIBRARY_PATH=$(SSL_LIBRARY_PATH)/lib
 endif
 
 ifndef TRAVIS
 # Which SSL implementation to use
 ifeq ($(SSL),)
-SSL = OPENSSL
+	SSL = OPENSSL
+	SSL_MODE = OPENSSL
 endif
 
 $(shell pkg-config --exists libssl libcrypto)
 ifeq ($(.SHELLSTATUS),0)
 ifeq ("$(SSL)","OPENSSL")
+	SSL_MODE = OPENSSL
 	FLAGS_EXTRA += $(shell pkg-config --libs libssl libcrypto)
 	CFLAGS += $(shell pkg-config --cflags libssl libcrypto) -DUSE_OPENSSL
 endif
@@ -114,6 +115,7 @@ endif
 $(shell pkg-config --exists wolfssl)
 ifeq ($(.SHELLSTATUS),0)
 ifeq ("$(SSL)","WOLFSSL")
+	SSL_MODE = WOLFSSL
 	FLAGS_EXTRA += $(shell pkg-config --libs wolfssl)
 	CFLAGS += $(shell pkg-config --cflags wolfssl) -DUSE_WOLFSSL
 endif
@@ -124,19 +126,20 @@ ifeq ($(.SHELLSTATUS),0)
 	FLAGS_TEST += $(shell pkg-config --libs criterion)
 endif
 else
+	SSL_MODE = OPENSSL
 	FLAGS_EXTRA += -lssl -lcrypto
 	CFLAGS += -DUSE_OPENSSL
 	FLAGS_TEST += -lcriterion
 endif
 
-
-.PHONY: valgrind version bump cachegrind callgrind clean subprotocols extensions autobahn massconnect autobahn_debug autobahn_call autobahn_cache analysis count release debug profiling space test ${addprefix run_,${TEST_NAMES}}
+.PHONY: valgrind version bump cachegrind callgrind clean subprotocols extensions autobahn massconnect massconnect_debug autobahn_debug autobahn_call autobahn_cache analysis count release debug profiling space test ${addprefix run_,${TEST_NAMES}}
 
 #what we are trying to build
-all: clean version bin build log subprotocols extensions $(NAME)
+all: internal_clean version bin build ssl log subprotocols extensions $(NAME)
 
 build:
 	if [[ ! -e $(BUILD_FOLDER) ]]; then mkdir -p $(BUILD_FOLDER); fi
+	echo "$(MODE)" >> $(BUILD_FOLDER)/.mode
 
 bin:
 	if [[ ! -e $(BIN_FOLDER) ]]; then mkdir -p $(BIN_FOLDER); fi
@@ -159,7 +162,7 @@ debug_mode:
 	$(eval MODE = "debug")
 
 profiling_mode:
-	$(eval EXEC = $(PROFILE))
+	$(eval EXEC = $(PROFILING))
 	$(eval MODE = "profiling")
 
 space_mode:
@@ -174,7 +177,7 @@ $(NAME): $(SRC_OBJ)
 	@echo 
 	@echo ================ [Linking] ================ 
 	@echo
-	$(CC) $(CFLAGS) ${FLAGS_COVERAGE} $(CVER) -o $(BIN_FOLDER)/$@ $(filter-out $(filter-out $(BUILD_FOLDER)/$@.o, $(addsuffix .o, $(addprefix $(BUILD_FOLDER)/, $(NAME)))), $^) $(FLAGS_EXTRA) $(INCLUDES)
+	$(CC) $(CFLAGS) ${FLAGS_COVERAGE} $(CVER) $(FLAGS_INCLUDES) -o $(BIN_FOLDER)/$@ $(filter-out $(filter-out $(BUILD_FOLDER)/$@.o, $(addsuffix .o, $(addprefix $(BUILD_FOLDER)/, $(NAME)))), $^) $(FLAGS_EXTRA)
 	@echo
 	@echo ================ [$(NAME) compiled succesfully] ================ 
 	@echo
@@ -184,7 +187,7 @@ $(BUILD_FOLDER)/%.o: $(SRC_FOLDER)/%.c
 	@echo
 	@echo ================ [Building Object] ================
 	@echo
-	$(CC) $(CFLAGS) ${FLAGS_COVERAGE} $(CVER) $(INCLUDES) -c $< -o $@
+	$(CC) $(CFLAGS) ${FLAGS_COVERAGE} $(CVER) $(FLAGS_INCLUDES) -c $< -o $@
 	@echo
 	@echo OK [$<] - [$@]
 	@echo
@@ -194,19 +197,19 @@ $(BUILD_FOLDER)/%.o: $(TEST_FOLDER)/%.c
 	@echo
 	@echo ================ [Building Object] ================
 	@echo
-	$(CC) --coverage $(CFLAGS) ${FLAGS_COVERAGE} $(CVER) $(INCLUDES) -c $< -o $@
+	$(CC) --coverage $(CFLAGS) ${FLAGS_COVERAGE} $(CVER) $(FLAGS_INCLUDES) -c $< -o $@
 	@echo
 	@echo OK [$<] - [$@]
 	@echo
 
 # Link test objects
-${TEST_NAMES}: clean debug_mode bin build log ${SRC_OBJ} ${TEST_OBJ}
+${TEST_NAMES}: internal_clean debug_mode bin build log ${SRC_OBJ} ${TEST_OBJ}
 	@echo
 	@echo ================ [Linking Tests] ================
 	@echo
-	$(CC) ${CFLAGS} ${FLAGS_COVERAGE} ${CVER} -o ${BIN_FOLDER}/$@ ${BUILD_FOLDER}/$@.o\
+	$(CC) ${CFLAGS} ${FLAGS_COVERAGE} ${CVER} $(FLAGS_INCLUDES) -o ${BIN_FOLDER}/$@ ${BUILD_FOLDER}/$@.o\
 		$(filter-out $(addsuffix .o, $(addprefix ${BUILD_FOLDER}/, main)), $(filter-out ${BUILD_FOLDER}/test_%.o, $(ALL_OBJ)))\
-		${FLAGS_EXTRA} ${FLAGS_TEST} $(INCLUDES)
+		${FLAGS_EXTRA} ${FLAGS_TEST}
 	@echo
 	@echo ================ [$@ compiled succesfully] ================
 
@@ -239,14 +242,18 @@ callgrind: profiling_mode all
 	@echo
 	valgrind --tool=callgrind --simulate-cache=yes --branch-sim=yes --callgrind-out-file=$(LOG_FOLDER)/$(NAME).callgrind.log $(BIN_FOLDER)/$(NAME) -c $(CONF_FOLDER)/wss.json
 
-#make clean
-clean:
+internal_clean:
 	@echo
 	@echo ================ [Cleaning $(NAME)] ================
 	@echo
+	if [ $(PREV_MODE) != $(MODE) ]; then rm -rf $(BUILD_FOLDER); fi
+	if [ "$(PREV_SSL)" != "$(SSL)" ]; then rm -rf $(BUILD_FOLDER); fi
 	rm -rf $(BIN_FOLDER)
-	rm -rf $(BUILD_FOLDER)
 	rm -rf $(LOG_FOLDER)
+
+#make clean
+clean: internal_clean
+	rm -rf $(BUILD_FOLDER);
 
 #make count
 count:
@@ -260,7 +267,7 @@ autobahn: release
 	rm -rf $(GEN_FOLDER)/autobahn
 	if [[ ! -e $(GEN_FOLDER) ]]; then mkdir -p $(GEN_FOLDER); fi
 	$(BIN_FOLDER)/$(NAME) -c $(CONF_FOLDER)/autobahn.json &
-	sleep 1
+	sleep 5
 	docker build -t wsserver/autobahn -f Dockerfile.test .
 	docker run -it --rm \
 	--network="host" \
@@ -277,7 +284,7 @@ autobahn_debug: debug
 	if [[ ! -e $(GEN_FOLDER) ]]; then mkdir -p $(GEN_FOLDER); fi
 	valgrind -v --leak-check=full --log-file="$(LOG_FOLDER)/valgrind.log" --track-origins=yes \
 	--show-reachable=yes $(BIN_FOLDER)/$(NAME) -c $(CONF_FOLDER)/autobahn.debug.json &
-	sleep 3
+	sleep 10
 	docker build -t wsserver/autobahn -f Dockerfile.test .
 	docker run -it --rm \
 	--network="host" \
@@ -323,8 +330,23 @@ autobahn_cache: profiling
 #make massconnect
 massconnect: release
 	if [[ ! -e $(GEN_FOLDER) ]]; then mkdir -p $(GEN_FOLDER); fi
-	$(BIN_FOLDER)/$(NAME) -c $(CONF_FOLDER)/autobahn.json &
+	$(BIN_FOLDER)/$(NAME) -c $(CONF_FOLDER)/autobahn-massconnect.json &
 	sleep 1
+	docker build -t wsserver/autobahn -f Dockerfile.massconnect .
+	docker run -it --rm \
+	--network="host" \
+    -v ${CONF_FOLDER}:/config \
+    -v ${GEN_FOLDER}:/generated \
+    -p 9001:9001 \
+    --name massconnect \
+    wsserver/autobahn
+	pkill $(NAME) || true
+
+#make massconnect_debug
+massconnect_debug: debug
+	if [[ ! -e $(GEN_FOLDER) ]]; then mkdir -p $(GEN_FOLDER); fi
+	$(BIN_FOLDER)/$(NAME) -c $(CONF_FOLDER)/autobahn-massconnect.json &
+	sleep 3
 	docker build -t wsserver/autobahn -f Dockerfile.massconnect .
 	docker run -it --rm \
 	--network="host" \
@@ -338,7 +360,7 @@ massconnect: release
 #make autobahn_cache
 massconnect_cache: profiling
 	if [[ ! -e $(GEN_FOLDER) ]]; then mkdir -p $(GEN_FOLDER); fi
-	valgrind --tool=cachegrind --trace-children=yes --cachegrind-out-file=$(LOG_FOLDER)/$(NAME).callgrind.log $(BIN_FOLDER)/$(NAME) -c $(CONF_FOLDER)/autobahn.json &
+	ulimit -n 100000 && valgrind --tool=cachegrind --trace-children=yes --cachegrind-out-file=$(LOG_FOLDER)/$(NAME).callgrind.log $(BIN_FOLDER)/$(NAME) -c $(CONF_FOLDER)/autobahn-massconnect.json &
 	sleep 3
 	docker build -t wsserver/autobahn -f Dockerfile.massconnect .
 	docker run -it --rm \
@@ -366,11 +388,13 @@ ${addprefix run_,${TEST_NAMES}}: ${TEST_NAMES}
 	${BIN_FOLDER}/${patsubst run_%,%,$@} --verbose
 
 analysis:
-	cppcheck --language=c -f -q --enable=warning,performance,portability --std=c11 --error-exitcode=1 -i$(TEST_FOLDER) $(INCLUDES) .
+	cppcheck --language=c -f -q --enable=warning,performance,portability -$(CVER) --error-exitcode=1 -i$(TEST_FOLDER) $(FLAGS_INCLUDES) .
  
 #make version
 version:
+	@echo
 	@echo ================ [$(NAME) - version $(VER)] ================
+	@echo
 
 #make release
 release: release_mode all
@@ -383,6 +407,9 @@ profiling: profiling_mode all
 
 #make space
 space: space_mode all
+
+ssl:
+	echo $(SSL_MODE) >> $(BUILD_FOLDER)/.ssl
 
 #make bump
 bump:
